+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Image 0.13.0
 
-This is the changelog for Image version 0.13.0 released on October 16th, 2022.  For older changelogs please consult the release tag on [GitHub](https://github.com/kipcole9/image/tags)
+This is the changelog for Image version 0.13.0 released on October ___th, 2022.  For older changelogs please consult the release tag on [GitHub](https://github.com/kipcole9/image/tags)
 
 ### Breaking change
 
@@ -15,6 +15,10 @@ This is the changelog for Image version 0.13.0 released on October 16th, 2022.  
 * Add `Image.resize/3`
 
 * Add `Image.blur/2`
+
+* Add `Image.blur/2`
+
+* Add `Image.preview/1` (and `Image.p/1` that delegates to it) to preview and image inline in iTerm2 terminal windows.
 
 ## Image 0.12.0